re-wrote bucket and finished bucket index construction.

+++ miniclean/common/types.h
@@ -23,7 +23,7 @@ typedef std::tuple<uint8_t, uint8_t> PatternVertexID;  // uint8_t: 0 ~ 255
 //   The vertex attribute ID specifies the number of bucket of the attribute.
 typedef uint8_t VertexAttributeID;
 
-typedef uint8_t VertexAttributeValue;
+typedef uint32_t VertexAttributeValue;
 
 // Since EdgeLabel is checked, VertexID is enough to represent an edge instance.
 typedef std::pair<VertexID, VertexID> EdgeInstance;
